Split project scoped Capnweb capabilities
